Add Claim conditions tests to Signature Drop
