+++ test/signature-mint-721-with-quantity-signature-minting.test.ts
@@ -1,10 +1,18 @@
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import { assert } from "chai";
-import { BigNumber } from "ethers";
-import { PayloadToSign721withQuantity, SignatureDrop, Token } from "../src";
-import { sdk, signers, storage } from "./before-setup";
+import { assert, expect } from "chai";
+import { BigNumber, ethers } from "ethers";
+import {
+  createSnapshot,
+  PayloadToSign721withQuantity,
+  SignatureDrop,
+  Token,
+} from "../src";
+import { expectError, sdk, signers, storage } from "./before-setup";
 import { SignedPayload721WithQuantitySignature } from "../src/schema/contracts/common/signature";
 import { NATIVE_TOKEN_ADDRESS } from "../src/constants/currency";
+import invariant from "tiny-invariant";
+import { MerkleTree } from "merkletreejs";
+import keccak256 from "keccak256";
 
 global.fetch = require("cross-fetch");
 
@@ -13,12 +21,19 @@ describe("ERC 721 with Signature minting", async () => {
   let customTokenContract: Token;
   let tokenAddress: string;
 
-  let adminWallet: SignerWithAddress, samWallet: SignerWithAddress;
+  let adminWallet: SignerWithAddress,
+    samWallet: SignerWithAddress,
+    abbyWallet: SignerWithAddress,
+    bobWallet: SignerWithAddress,
+    w1: SignerWithAddress,
+    w2: SignerWithAddress,
+    w3: SignerWithAddress,
+    w4: SignerWithAddress;
 
   let meta: PayloadToSign721withQuantity;
 
   before(() => {
-    [adminWallet, samWallet] = signers;
+    [adminWallet, samWallet, bobWallet, abbyWallet, w1, w2, w3, w4] = signers;
   });
 
   beforeEach(async () => {
@@ -168,30 +183,6 @@ describe("ERC 721 with Signature minting", async () => {
       ]);
     });
 
-    // it("should allow batch minting", async () => {
-    //   const payloads = [];
-    //   const freeMint = {
-    //     currencyAddress: NATIVE_TOKEN_ADDRESS,
-    //     metadata: {
-    //       name: "OUCH VOUCH",
-    //     },
-    //     price: 0,
-    //     quantity: 1,
-    //     to: samWallet.address,
-    //   };
-    //   for (let i = 0; i < 10; i++) {
-    //     payloads.push(freeMint);
-    //   }
-    //   const batch = await Promise.all(
-    //     payloads.map(async (p) => await signatureDropContract.signature.generate(p)),
-    //   );
-    //   await sdk.updateSignerOrProvider(samWallet);
-    //   const tx = await signatureDropContract.signature.mintBatch(batch);
-    //   expect(tx.length).to.eq(10);
-    //   expect(tx[0].id.toNumber()).to.eq(0);
-    //   expect(tx[3].id.toNumber()).to.eq(3);
-    // });
-
     it("should mint with URI", async () => {
       const uri = await storage.uploadMetadata({
         name: "Test1",
@@ -254,40 +245,6 @@ describe("ERC 721 with Signature minting", async () => {
       assert.equal(nft2.metadata.name, "Test2");
     });
 
-    // it("should mint additional supply", async () => {
-    //   const tx = await signatureDropContract.signature.mint(v1);
-    //   const additional = await signatureDropContract.signature.generate({
-    //     tokenId: tx.id,
-    //     quantity: 100,
-    //   });
-    //   await signatureDropContract.signature.mint(additional);
-    //   const nft = await signatureDropContract.get(tx.id);
-    //   expect(nft.supply.toNumber()).to.eq(101);
-    // });
-
-    // it("should mint additional supply of one tokenId", async () => {
-    //   const oldBalance = await signatureDropContract.balanceOf(
-    //     samWallet.address,
-    //     "0",
-    //   );
-    //   await signatureDropContract.mintToSelf({
-    //     metadata: { name: "test" },
-    //     supply: 0,
-    //   });
-    //   const payload = await signatureDropContract.signature.generate({
-    //     tokenId: "0",
-    //     quantity: "1",
-    //     metadata: "",
-    //   });
-    //   sdk.updateSignerOrProvider(samWallet);
-    //   await signatureDropContract.signature.mint(payload);
-    //   const newBalance = await signatureDropContract.balanceOf(
-    //     samWallet.address,
-    //     "0",
-    //   );
-    //   assert(newBalance.gt(oldBalance), "balance doesn't match");
-    // });
-
     it("should mint the right custom token price", async () => {
       const oldBalance = await samWallet.getBalance();
       const payload = await signatureDropContract.signature.generate({
@@ -347,6 +304,483 @@ describe("ERC 721 with Signature minting", async () => {
         oldBalance.sub(newBalance).gte(BigNumber.from(2)),
         "balance doesn't match",
       );
+    });
+  });
+
+  describe("Delay Reveal", () => {
+    it("metadata should reveal correctly", async () => {
+      await signatureDropContract.revealer.createDelayedRevealBatch(
+        {
+          name: "Placeholder #1",
+        },
+        [{ name: "NFT #1" }, { name: "NFT #2" }],
+        "my secret password",
+      );
+
+      expect((await signatureDropContract.get("0")).metadata.name).to.be.equal(
+        "Placeholder #1",
+      );
+
+      await signatureDropContract.revealer.reveal(0, "my secret password");
+
+      expect((await signatureDropContract.get("0")).metadata.name).to.be.equal(
+        "NFT #1",
+      );
+    });
+
+    it("different reveal order and should return correct unreveal list", async () => {
+      await signatureDropContract.revealer.createDelayedRevealBatch(
+        {
+          name: "Placeholder #1",
+        },
+        [
+          {
+            name: "NFT #1",
+          },
+          {
+            name: "NFT #2",
+          },
+        ],
+        "my secret key",
+      );
+
+      await signatureDropContract.revealer.createDelayedRevealBatch(
+        {
+          name: "Placeholder #2",
+        },
+        [
+          {
+            name: "NFT #3",
+          },
+          {
+            name: "NFT #4",
+          },
+        ],
+        "my secret key",
+      );
+
+      await signatureDropContract.createBatch([
+        {
+          name: "NFT #00",
+        },
+        {
+          name: "NFT #01",
+        },
+      ]);
+
+      await signatureDropContract.revealer.createDelayedRevealBatch(
+        {
+          name: "Placeholder #3",
+        },
+        [
+          {
+            name: "NFT #5",
+          },
+          {
+            name: "NFT #6",
+          },
+          {
+            name: "NFT #7",
+          },
+        ],
+        "my secret key",
+      );
+
+      let unrevealList =
+        await signatureDropContract.revealer.getBatchesToReveal();
+      expect(unrevealList.length).to.be.equal(3);
+      expect(unrevealList[0].batchId.toNumber()).to.be.equal(0);
+      expect(unrevealList[0].placeholderMetadata.name).to.be.equal(
+        "Placeholder #1",
+      );
+      expect(unrevealList[1].batchId.toNumber()).to.be.equal(1);
+      expect(unrevealList[1].placeholderMetadata.name).to.be.equal(
+        "Placeholder #2",
+      );
+      // skipped 2 because it is a revealed batch
+      expect(unrevealList[2].batchId.toNumber()).to.be.equal(3);
+      expect(unrevealList[2].placeholderMetadata.name).to.be.equal(
+        "Placeholder #3",
+      );
+
+      await signatureDropContract.revealer.reveal(
+        unrevealList[0].batchId,
+        "my secret key",
+      );
+
+      unrevealList = await signatureDropContract.revealer.getBatchesToReveal();
+      expect(unrevealList.length).to.be.equal(2);
+      expect(unrevealList[0].batchId.toNumber()).to.be.equal(1);
+      expect(unrevealList[0].placeholderMetadata.name).to.be.equal(
+        "Placeholder #2",
+      );
+      expect(unrevealList[1].batchId.toNumber()).to.be.equal(3);
+      expect(unrevealList[1].placeholderMetadata.name).to.be.equal(
+        "Placeholder #3",
+      );
+
+      await signatureDropContract.revealer.reveal(
+        unrevealList[unrevealList.length - 1].batchId,
+        "my secret key",
+      );
+
+      unrevealList = await signatureDropContract.revealer.getBatchesToReveal();
+      expect(unrevealList.length).to.be.equal(1);
+      expect(unrevealList[0].batchId.toNumber()).to.be.equal(1);
+      expect(unrevealList[0].placeholderMetadata.name).to.be.equal(
+        "Placeholder #2",
+      );
+    });
+
+    it("should not be able to re-used published password for next batch", async () => {
+      await signatureDropContract.revealer.createDelayedRevealBatch(
+        {
+          name: "Placeholder #1",
+        },
+        [{ name: "NFT #1" }, { name: "NFT #2" }],
+        "my secret password",
+      );
+      await signatureDropContract.revealer.createDelayedRevealBatch(
+        {
+          name: "Placeholder #2",
+        },
+        [{ name: "NFT #3" }, { name: "NFT #4" }],
+        "my secret password",
+      );
+      await signatureDropContract.revealer.reveal(0, "my secret password");
+      const transactions =
+        (await adminWallet.provider?.getBlockWithTransactions("latest"))
+          ?.transactions ?? [];
+
+      const { _index, _key } = signatureDropContract.encoder.decode(
+        "reveal",
+        transactions[0].data,
+      );
+
+      // re-using broadcasted _key to decode :)
+      try {
+        await signatureDropContract.revealer.reveal(_index.add(1), _key);
+        assert.fail("should not be able to re-used published password");
+      } catch (e) {
+        expect((e as Error).message).to.be.equal("invalid password");
+      }
+
+      // original password should work
+      await signatureDropContract.revealer.reveal(1, "my secret password");
+    });
+  });
+
+  describe("Claim Conditions", () => {
+    it("sshould allow a snapshot to be set", async () => {
+      await signatureDropContract.createBatch([
+        { name: "test", description: "test" },
+        { name: "test", description: "test" },
+      ]);
+
+      await signatureDropContract.claimConditions.set([
+        {
+          snapshot: [samWallet.address],
+        },
+      ]);
+      const conditions = await signatureDropContract.claimConditions.getAll();
+      assert.lengthOf(conditions, 1);
+      invariant(conditions[0].snapshot);
+      expect(conditions[0].snapshot[0].address).to.eq(samWallet.address);
+    });
+
+    it("should remove merkles from the metadata when claim conditions are removed", async () => {
+      await signatureDropContract.claimConditions.set([
+        {
+          startTime: new Date(),
+          waitInSeconds: 10,
+          snapshot: [bobWallet.address, samWallet.address, abbyWallet.address],
+        },
+        {
+          startTime: new Date(Date.now() + 60 * 60 * 1000),
+          snapshot: [bobWallet.address],
+        },
+      ]);
+
+      const metadata = await signatureDropContract.metadata.get();
+      const merkles = metadata.merkle;
+
+      expect(merkles).have.property(
+        "0xd89eb21bf7ee4dd07d88e8f90a513812d9d38ac390a58722762c9f3afc4e0feb",
+      );
+
+      expect(merkles).have.property(
+        "0xb1a60ad68b77609a455696695fbdd02b850d03ec285e7fe1f4c4093797457b24",
+      );
+
+      const roots = (await signatureDropContract.claimConditions.getAll()).map(
+        (c) => c.merkleRootHash,
+      );
+      expect(roots).length(2);
+
+      await signatureDropContract.claimConditions.set([{}]);
+      const newMetadata = await signatureDropContract.metadata.get();
+      const newMerkles = newMetadata.merkle;
+      expect(JSON.stringify(newMerkles)).to.eq("{}");
+    });
+
+    it("allow all addresses in the merkle tree to claim", async () => {
+      const testWallets: SignerWithAddress[] = [
+        bobWallet,
+        samWallet,
+        abbyWallet,
+        w1,
+        w2,
+        w3,
+        w4,
+      ];
+      const members = testWallets.map((w, i) =>
+        i % 3 === 0
+          ? w.address.toLowerCase()
+          : i % 3 === 1
+          ? w.address.toUpperCase().replace("0X", "0x")
+          : w.address,
+      );
+
+      await signatureDropContract.claimConditions.set([
+        {
+          snapshot: members,
+        },
+      ]);
+      const metadata = [];
+      for (let i = 0; i < 10; i++) {
+        metadata.push({
+          name: `test ${i}`,
+        });
+      }
+      await signatureDropContract.createBatch(metadata);
+
+      /**
+       * Claiming 1 tokens with proofs: 0xe9707d0e6171f728f7473c24cc0432a9b07eaaf1efed6a137a4a8c12c79552d9,0xb1a5bda84b83f7f014abcf0cf69cab5a4de1c3ececa8123a5e4aaacb01f63f83
+       */
+
+      for (const member of testWallets) {
+        await sdk.updateSignerOrProvider(member);
+        await signatureDropContract.claim(1);
+      }
+    });
+
+    it("allow one address in the merkle tree to claim", async () => {
+      const testWallets: SignerWithAddress[] = [bobWallet];
+      const members = testWallets.map((w) => w.address);
+
+      await signatureDropContract.claimConditions.set([
+        {
+          snapshot: members,
+        },
+      ]);
+
+      const metadata = [];
+      for (let i = 0; i < 2; i++) {
+        metadata.push({
+          name: `test ${i}`,
+        });
+      }
+      await signatureDropContract.createBatch(metadata);
+
+      /**
+       * Claiming 1 tokens with proofs: 0xe9707d0e6171f728f7473c24cc0432a9b07eaaf1efed6a137a4a8c12c79552d9,0xb1a5bda84b83f7f014abcf0cf69cab5a4de1c3ececa8123a5e4aaacb01f63f83
+       */
+
+      for (const member of testWallets) {
+        await sdk.updateSignerOrProvider(member);
+        await signatureDropContract.claim(1);
+      }
+
+      try {
+        await sdk.updateSignerOrProvider(samWallet);
+        await signatureDropContract.claim(1);
+        assert.fail("should have thrown");
+      } catch (e) {
+        // expected
+      }
+    });
+
+    it("should correctly upload metadata for each nft", async () => {
+      const metadatas = [];
+      for (let i = 0; i < 100; i++) {
+        metadatas.push({
+          name: `test ${i}`,
+        });
+      }
+      await signatureDropContract.createBatch(metadatas);
+      const all = await signatureDropContract.getAll();
+      expect(all.length).to.eq(100);
+      await signatureDropContract.claimConditions.set([{}]);
+      await signatureDropContract.claim(1);
+      const claimed = await signatureDropContract.getAllClaimed();
+      const unclaimed = await signatureDropContract.getAllUnclaimed({
+        start: 0,
+        count: 30,
+      });
+      expect(claimed.length).to.eq(1);
+      expect(unclaimed.length).to.eq(30);
+      expect(unclaimed[0].name).to.eq("test 1");
+      expect(unclaimed[unclaimed.length - 1].name).to.eq("test 30");
+    });
+
+    it("should correctly update total supply after burning", async () => {
+      const metadatas = [];
+      for (let i = 0; i < 20; i++) {
+        metadatas.push({
+          name: `test ${i}`,
+        });
+      }
+      await signatureDropContract.createBatch(metadatas);
+      await signatureDropContract.claimConditions.set([{}]);
+      await signatureDropContract.claim(10);
+      const ts = await signatureDropContract.totalSupply();
+      expect(ts.toNumber()).to.eq(20);
+      await signatureDropContract.burn(0);
+      const ts2 = await signatureDropContract.totalSupply();
+      expect(ts2.toNumber()).to.eq(20);
+    });
+
+    it("should not allow claiming to someone not in the merkle tree", async () => {
+      await signatureDropContract.claimConditions.set(
+        [
+          {
+            snapshot: [
+              bobWallet.address,
+              samWallet.address,
+              abbyWallet.address,
+            ],
+          },
+        ],
+        false,
+      );
+      await signatureDropContract.createBatch([
+        { name: "name", description: "description" },
+      ]);
+
+      await sdk.updateSignerOrProvider(w1);
+      try {
+        await signatureDropContract.claim(1);
+      } catch (err: any) {
+        expect(err).to.have.property(
+          "message",
+          "No claim found for this address",
+          "",
+        );
+        return;
+      }
+      assert.fail("should not reach this point, claim should have failed");
+    });
+
+    it("should allow claims with default settings", async () => {
+      await signatureDropContract.createBatch([
+        { name: "name", description: "description" },
+      ]);
+      await signatureDropContract.claimConditions.set([{}]);
+      await signatureDropContract.claim(1);
+    });
+
+    it("should allow setting max claims per wallet", async () => {
+      await signatureDropContract.createBatch([
+        { name: "name", description: "description" },
+        { name: "name2", description: "description" },
+        { name: "name3", description: "description" },
+        { name: "name4", description: "description" },
+      ]);
+      await signatureDropContract.claimConditions.set([
+        {
+          snapshot: [
+            { address: w1.address, maxClaimable: 2 },
+            { address: w2.address, maxClaimable: 1 },
+          ],
+        },
+      ]);
+      await sdk.updateSignerOrProvider(w1);
+      const tx = await signatureDropContract.claim(2);
+      expect(tx.length).to.eq(2);
+      try {
+        await sdk.updateSignerOrProvider(w2);
+        await signatureDropContract.claim(2);
+      } catch (e) {
+        expectError(e, "invalid quantity proof");
+      }
+    });
+
+    it("should generate valid proofs", async () => {
+      const members = [
+        bobWallet.address,
+        samWallet.address,
+        abbyWallet.address,
+        w1.address,
+        w2.address,
+        w3.address,
+        w4.address,
+      ];
+
+      const hashedLeafs = members.map((l) =>
+        ethers.utils.solidityKeccak256(["address", "uint256"], [l, 0]),
+      );
+      const tree = new MerkleTree(hashedLeafs, keccak256, {
+        sort: true,
+        sortLeaves: true,
+        sortPairs: true,
+      });
+      const input = members.map((address) => ({
+        address,
+        maxClaimable: 0,
+      }));
+      const snapshot = await createSnapshot(input, 0, storage);
+      for (const leaf of members) {
+        const expectedProof = tree.getHexProof(
+          ethers.utils.solidityKeccak256(["address", "uint256"], [leaf, 0]),
+        );
+
+        const actualProof = snapshot.snapshot.claims.find(
+          (c) => c.address === leaf,
+        );
+        assert.isDefined(actualProof);
+        expect(actualProof?.proof).to.include.ordered.members(expectedProof);
+
+        const verified = tree.verify(
+          actualProof?.proof as string[],
+          ethers.utils.solidityKeccak256(["address", "uint256"], [leaf, 0]),
+          tree.getHexRoot(),
+        );
+        expect(verified).to.eq(true);
+      }
+    });
+
+    it("should return the newly claimed token", async () => {
+      await signatureDropContract.claimConditions.set([{}]);
+      await signatureDropContract.createBatch([
+        {
+          name: "test 0",
+        },
+        {
+          name: "test 1",
+        },
+        {
+          name: "test 2",
+        },
+      ]);
+
+      try {
+        await signatureDropContract.createBatch([
+          {
+            name: "test 0",
+          },
+          {
+            name: "test 1",
+          },
+          {
+            name: "test 2",
+          },
+        ]);
+      } catch (err) {
+        expect(err).to.have.property("message", "Batch already created!", "");
+      }
+
+      const token = await signatureDropContract.claim(2);
+      assert.lengthOf(token, 2);
     });
   });
 });
